refactor: replace shm_wp_list -> wp_list used shared_ptr
