+++ src/shm_board.h
@@ -4,54 +4,17 @@
 #include <iostream>
 #include <cstring>
 #include <vector>
-#include <sys/shm.h>
+//#include <sys/shm.h>
 #include <stdlib.h>
 #include <string_view>
-#include <semaphore.h>
-#include "Urg2d.h"
+//#include <semaphore.h>
+//#include "Urg2d.h"
 
-
-// shm取得から割り当てをまとめて行う
-void *shmAt(key_t KEY, size_t key_size) {
-	int keyID = shmget(KEY, key_size, 0666 | IPC_CREAT);
-  if (keyID == -1) {
-    std::cerr << KEY << "のshmgetに失敗しました\n";
-    std::cerr << std::strerror(errno) << "\n";
-    exit(1);
-  } else {
-    //std::cerr << keyID << "を取得済み\n";
-  }
-  return shmat(keyID, 0, 0);
-}
-
-// for ROUTE_LIST
-#define KEY_WP_LIST 1288
-enum class ChangeWPTrigger {
-	kContinue = 0,
-	kChange = 1,
-};
-struct ROUTE_POINT {
-	double x;
-	double y;
-	double a;
-  int stop_check;
-};
-struct WP_LIST {
-	char path_to_wp_file[256]; 			// 現在使用中のWPファイルへのフルパス
-	ChangeWPTrigger change_wp_trigger;	// WPのリセットトリガー
-	int size_wp_list;
-	int size_route_list;
-	int target_index;
-	ROUTE_POINT wp_list[3000]; 		// WAY POINT
-	ROUTE_POINT route_list[3000]; 	// 細分化した通過点
-	int wp_index_list[3000]; 		// 通過点が目指しているWAY POINTのインデックス
-};
 
 // for LOG Display
 const int LOG_HEIGHT = 10;
 #define LOG_SIZE 1024
 #define NUM_LOGS 1000
-// 共有メモリのデータ構造
 struct LOG_DATA {
     size_t current_index; // 現在のログインデックス
     std::vector<std::string> logs;
@@ -61,6 +24,7 @@ struct LOG_DATA {
       logs.resize(NUM_LOGS);
     }
 };
+
 // ログの追加
 void add_log(std::shared_ptr<LOG_DATA> shared, std::string_view log) {
   {
